# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentación de la API generada automáticamente con Swagger",
        "contact": {
            "email": "person@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/roles/index": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Obtener todos los roles",
                "operationId": "92e30fe1eb2346d6e8e4b9ac21208359",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de roles por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Role"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "total_pages": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string"
                                                },
                                                "last": {
                                                    "type": "string"
                                                },
                                                "prev": {
                                                    "type": "string"
                                                },
                                                "next": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Parámetros incorrectos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/roles/create": {
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Crear un nuevo rol",
                "operationId": "6106e5a4a614ac06977b5d4de4689052",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rol creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/roles/show/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Obtener un rol por ID",
                "operationId": "77deefcb47aad4da253b2d61349865e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del rol",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/roles/update/{role}": {
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Actualizar un rol existente",
                "operationId": "bb668def592b2dabfff17b0f17fd065c",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rol actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/roles/delete/{role}": {
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Eliminar un rol",
                "operationId": "90f8e8ea85c4659a48df9fafd5184d50",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Role deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Role": {
                "title": "Role",
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "ID del rol",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre del rol",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "RoleRequest": {
                "title": "Role Request",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre del rol",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header",
                "description": "Token JWT necesario para acceder a la API"
            }
        }
    },
...
}